requirements.txt
psycopg2-binary==2.9.9
playwright==1.40.0
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
